+++ android/src/main/java/com/qliktrialreactnativestraighttable/DataProvider.java
@@ -217,9 +217,6 @@ public class DataProvider extends RecyclerView.Adapter<RecyclerView.ViewHolder> 
       }
       this.notifyDataSetChanged();
     }
-    if(!isDataView) {
-      tableView.imageLoader.loadImages();
-    }
     setLoading(false);
   }
   public void setDataColumns(List<DataColumn> cols) {
@@ -233,7 +230,7 @@ public class DataProvider extends RecyclerView.Adapter<RecyclerView.ViewHolder> 
         .filter(dataCol -> dataCol.dataColIdx == col.dataColIdx)
         .findAny()
         .orElse(col);
-      col.width = column.width == 0 ? columnWidths.resizeColumnByAverage(column, rows, true) : column.width;
+      col.width = column.width == 0 ? columnWidths.resizeColumnByAverage(column, rows, true, tableView.totalWidth) : column.width;
     }).collect(Collectors.toList());
   }